restructure project sources
