+++ packages/roaring-wasm-src/cpp/global.h
